Refactor functions in modules - part 1
